chatbot/requirements.txt
fastapi
uvicorn
pymongo
motor
pydantic
httpx
python-dotenv
streamlit
